requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
altair==5.5.0
altex==0.2.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
asn1crypto==1.5.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.2
bleach==6.3.0
blinker==1.9.0
blis==1.3.0
boto3==1.40.59
botocore==1.40.59
cachetools==6.2.1
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
cloudpickle==3.1.1
colorama==0.4.6
confection==0.1.5
contourpy==1.3.2
cryptography==46.0.3
cssselect==1.3.0
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
Django==5.2.7
djangorestframework==3.16.1
docopt==0.6.2
docx2txt==0.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
entrypoints==0.4
exceptiongroup==1.3.0
executing==2.2.1
faiss-cpu==1.12.0
Faker==37.12.0
fastjsonschema==2.21.2
favicon==0.7.0
filelock==3.20.0
Flask==3.1.2
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
htbuilder==0.9.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
idna==3.11
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
jwt==1.4.0
kiwisolver==1.4.9
langchain==1.0.0
langchain-classic==1.0.0
langchain-community==0.4
langchain-core==1.0.0
langchain-huggingface==1.0.0
langchain-text-splitters==1.0.0
langcodes==3.5.0
langgraph==1.0.0
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.0
langgraph-sdk==0.2.9
langsmith==0.4.37
language_data==1.3.0
lxml==6.0.2
lxml_html_clean==0.4.3
marisa-trie==1.3.1
Markdown==3.9
markdown-it-py==4.0.0
markdownlit==0.0.7
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
mpmath==1.3.0
multidict==6.7.0
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==2.8.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.4.2
nltk==3.9.2
numpy==1.26.4
oauthlib==3.3.1
odfpy==1.4.1
openai==2.7.1
orjson==3.11.3
ormsgpack==1.11.0
packaging==24.2
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
passlib==1.7.4
pdf2image==1.17.0
pickleshare==0.7.5
pillow==11.3.0
pipreqs==0.5.0
platformdirs==4.5.0
plotly==6.3.1
preshed==3.0.10
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pymdown-extensions==10.16.1
PyMuPDF==1.26.5
pyOpenSSL==25.3.0
pyparsing==3.2.5
PyPDF2==3.0.1
pyphen==0.17.2
pyspellchecker==0.8.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
readability-lxml==0.8.4.1
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.1.1
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==4.0.0
snowflake-snowpark-python==1.41.0
sortedcontainers==2.4.0
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.35
sqlparse==0.5.3
srsly==2.5.1
st-annotated-text==4.0.2
st-theme==1.2.3
stack-data==0.6.3
streamlit==1.51.0
streamlit-autorefresh==1.0.1
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.8
streamlit-image-coordinates==0.4.0
streamlit-keyup==0.3.0
streamlit-notify==0.3.1
streamlit-option-menu==0.4.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_faker==0.0.4
sympy==1.14.0
tenacity==9.1.2
textstat==0.7.10
thinc==8.3.6
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
tomlkit==0.13.3
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
vaderSentiment==3.3.2
validators==0.35.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.14
weasel==0.4.1
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.3
xxhash==3.6.0
yarg==0.1.9
yarl==1.22.0
zstandard==0.25.0
